Refractor header creation.

+++ graphic-file-converter/Utils.cpp
@@ -15,7 +15,7 @@ unsigned Utils::fourCharsToInt(const char* input, const int offset)
 	unsigned int out = 0;
 	for (int i = 0; i < 4; ++i)
 	{
-		out |= input[offset + i] << 8 * i;
+		out |= static_cast<uint8_t>(input[offset + i]) << 8 * i;
 	}
 	return out;
 }
